Create new stylesheet after renaming component to fontResize.js
