+++ src/shared/table-options/FontResize.js
